Moved parseRepType to helpers
